New Camera because of qFuzzyCompare

+++ flatterraingenerator.cpp
@@ -22,7 +22,7 @@ public:
 private:
     Qt3DExtras::QPlaneGeometry *mTileGeometry;
 
-    int const basePlaneDimesion = 65024;
+    float const basePlaneDimesion = 65024.0;
 };
 
 
@@ -58,9 +58,9 @@ Qt3DCore::QEntity *FlatTerrainChunkLoader::createEntity(Qt3DCore::QEntity *paren
     transform = new Qt3DCore::QTransform();
     entity->addComponent(transform);
 
-    int size = basePlaneDimesion / qPow(2, tile->z());
-    int xOffset = (tile->x() + 0.5) * size - basePlaneDimesion / 2;
-    int yOffset = (tile->y() + 0.5) * size - basePlaneDimesion / 2;
+    float size = basePlaneDimesion / qPow(2, tile->z());
+    float xOffset = (tile->x() + 0.5) * size - basePlaneDimesion / 2;
+    float yOffset = (tile->y() + 0.5) * size - basePlaneDimesion / 2;
 
     transform->setScale(size);
     transform->setTranslation(QVector3D(xOffset, 0, yOffset));
